Added separate turret logic for shootingController
