+++ Assets/Scripts/Dynamics/TurretIK.cs
@@ -22,21 +22,19 @@ public class TurretIK : MonoBehaviour
     private Vector3 CountAxis(Vector3 rot, Vector3 fac)
     {
         return new Vector3(rot.x * fac.x, rot.y * fac.y, rot.z * fac.z);
-    } 
+    }
 
-    private void RotateTurret()
+    private void RotateTurret(Vector3 position)
     {
-        Vector3 rotateTo = Quaternion.LookRotation(target.transform.position - transform.position).eulerAngles;
+        Vector3 rotateTo = Quaternion.LookRotation(position - transform.position).eulerAngles;
         transform.rotation = Quaternion.LerpUnclamped(transform.rotation, Quaternion.Euler(initialEulers + CountAxis(rotateTo, rotationAxis)), Time.deltaTime * rotationSpeed);
         gunBone.localRotation = Quaternion.Euler(gunInitialEulers + CountAxis(rotateTo, gunRotationAxis));
     }
 
     private void Update()
     {
-        if (target)
-        {
-            RotateTurret();
-        }
+        if (target) 
+            RotateTurret(target.position);
     }
 
 #if UNITY_EDITOR
