+++ Assets/Scripts/Units/Base Classes/Controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be60e3a0827c0bc47950b050cd8ccf53
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
