+++ Assets/Scripts/Units/Base Classes/Controller/TurretController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29d36a99f79c93f4282710c6755143fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
