+++ Assets/Scripts/Units/Base Classes/TurretMechanisedUnitBase.cs
@@ -5,18 +5,8 @@ using UnityEngine;
 
 public class TurretMechanisedUnitBase : ShootingUnitBase, IUnit, IShooting, IDamagable, IMechanisedUnit
 {
-    [Header("Turret Logic")]
-    [SerializeField, ChildGameObjectsOnly] private TurretIK turretController;
-
     public override void OnDeath()
     {
-        // Do smth
-    }
-
-    protected override Transform FindClosest(Teams team)
-    {
-        Transform closest = base.FindClosest(team);
-        turretController.Target = closest;
-        return closest;
+        // Do smth.
     }
 }
